+++ src/Forcing/Forcing.jl
@@ -1,0 +1,10 @@
+module Forcing
+
+export ModelForcing, SimpleForcing
+
+zeroforcing(args...) = 0
+
+include("simple_forcing.jl")
+include("model_forcing.jl")
+
+end
