+++ src/Oceananigans.jl
@@ -216,7 +216,7 @@ include("Solvers/Solvers.jl")
 
 using .Solvers
 
-include("forcing.jl")
+include("Forcing/Forcing.jl")
 include("logger.jl")
 include("Models/Models.jl")
